+++ packages/date-picker/src/DateRangePicker/DateRangePicker.tsx
@@ -56,7 +56,7 @@ export interface DateRangePickerProps extends DisabledDayMatchers {
 export const DateRangePicker = ({
   id,
   buttonRef = useRef<HTMLButtonElement>(null),
-  description,
+  description: _description, // not used
   labelText,
   isDisabled = false,
   classNameOverride,
@@ -141,7 +141,7 @@ export const DateRangePicker = ({
 
   return (
     <div>
-      <div ref={containerRef}>
+      <div ref={containerRef} className={classNameOverride}>
         <Label disabled={isDisabled} htmlFor={id} labelText={labelText} />
         <button
           id={id}