+++ packages/design-tokens/src/lib/__tests__/mapLeafsOfObject.spec.ts
@@ -1,8 +1,8 @@
 import { objectPathToCssVarFunction } from "../cssVariables"
 import { mapLeafsOfObject } from "../mapLeafsOfObject"
 
-describe(mapLeafsOfObject.name, () => {
-  test.each([
+describe("mapLeafsOfObject()", () => {
+  it.each([
     [
       {
         test: {