+++ draft-packages/collapsible/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.28](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@3.5.27...@kaizen/draft-collapsible@3.5.28) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-collapsible
+
+
+
+
+
+## [3.5.27](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@3.5.26...@kaizen/draft-collapsible@3.5.27) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-collapsible
+
+
+
+
+
 ## [3.5.26](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@3.5.25...@kaizen/draft-collapsible@3.5.26) (2023-03-23)
 
 **Note:** Version bump only for package @kaizen/draft-collapsible