+++ packages/split-button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.42](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/split-button@1.3.41...@kaizen/split-button@1.3.42) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/split-button
+
+
+
+
+
+## [1.3.41](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/split-button@1.3.40...@kaizen/split-button@1.3.41) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/split-button
+
+
+
+
+
 ## [1.3.40](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/split-button@1.3.39...@kaizen/split-button@1.3.40) (2023-03-23)
 
 **Note:** Version bump only for package @kaizen/split-button